Move nfiq2_version.h
