+++ NFIQ2/NFIQ2Algorithm/include/nfiq2/version.h
