+++ NFIQ2/NFIQ2Algorithm/src/nfiq2/version.cpp.in
@@ -4,7 +4,7 @@
 #include "FRFXLL.h"
 #include "opencv2/core/version.hpp"
 
-#include "nfiq2_version.h"
+#include "nfiq2/version.h"
 
 const unsigned int NFIQ::Version::Major = @NFIQ2_VERSION_MAJOR@;
 const unsigned int NFIQ::Version::Minor = @NFIQ2_VERSION_MINOR@;
