+++ NFIQ2/NFIQ2Algorithm/src/tool/nfiq2_ui_utils.cpp
@@ -9,8 +9,8 @@
  ******************************************************************************/
 
 #include <be_io_utility.h>
-#include <nfiq2_version.h>
 
+#include "nfiq2/version.h"
 #include "nfiq2_ui_log.h"
 #include "nfiq2_ui_types.h"
 #include "nfiq2_ui_utils.h"
